+++ deepmd/pt/model/model/dos_model.py
@@ -12,14 +12,14 @@ from deepmd.pt.model.atomic_model import (
 from deepmd.pt.model.model.model import (
     BaseModel,
 )
-
+from .dp_model import DPModel
 from .make_model import (
     make_model,
 )
 
 
 @BaseModel.register("standard")
-class DOSModel(make_model(DPAtomicModel)):
+class DOSModel(make_model(DPAtomicModel), DPModel):
     model_type = "dos"
 
     def __init__(
@@ -57,23 +57,6 @@ class DOSModel(make_model(DPAtomicModel)):
             model_predict = model_ret
             model_predict["updated_coord"] += coord
         return model_predict
-
-    @classmethod
-    def update_sel(cls, global_jdata: dict, local_jdata: dict):
-        """Update the selection and perform neighbor statistics.
-
-        Parameters
-        ----------
-        global_jdata : dict
-            The global data, containing the training section
-        local_jdata : dict
-            The local data refer to the current class
-        """
-        local_jdata_cpy = local_jdata.copy()
-        local_jdata_cpy["descriptor"] = cls.get_descriptor().update_sel(
-            global_jdata, local_jdata["descriptor"]
-        )
-        return local_jdata_cpy
 
     def get_fitting_net(self):
         """Get the fitting network."""
