+++ deepmd/pt/model/model/spin_model.py
@@ -18,8 +18,11 @@ from deepmd.utils.spin import (
     Spin,
 )
 
-from .dp_model import (
-    DPModel,
+from .make_model import (
+    make_model,
+)
+from deepmd.pt.model.atomic_model import (
+    DPAtomicModel,
 )
 
 
@@ -474,7 +477,7 @@ class SpinModel(torch.nn.Module):
 
     @classmethod
     def deserialize(cls, data) -> "SpinModel":
-        backbone_model_obj = DPModel.deserialize(data["backbone_model"])
+        backbone_model_obj = make_model(DPAtomicModel).deserialize(data["backbone_model"])
         spin = Spin.deserialize(data["spin"])
         return cls(
             backbone_model=backbone_model_obj,
